Refactor ReplitEmbed
